test(P_IMPLEMENTATION_SP_C6_RED): Move GitHub read models to github_models.py (D6.1) (#402)
