+++ .phase-gate/state.json
@@ -3,8 +3,8 @@
   "issue_number": 402,
   "workflow_name": "feature",
   "current_phase": "implementation",
-  "current_cycle": 5,
-  "last_cycle": 4,
+  "current_cycle": 6,
+  "last_cycle": 5,
   "cycle_history": [
     {
       "cycle_number": 8,
@@ -108,6 +108,12 @@
       "name": "Batch 3b (Git Mutation & Action Tools)",
       "forced": false,
       "entered": "2026-06-14T17:38:00.075130+00:00"
+    },
+    {
+      "cycle_number": 6,
+      "name": "Batch 4a (GitHub Issues & PRs)",
+      "forced": false,
+      "entered": "2026-06-14T19:17:51.355188+00:00"
     }
   ],
   "required_phases": [
@@ -295,5 +301,5 @@
     }
   ],
   "reconstructed": false,
-  "current_sub_phase": "red"
+  "current_sub_phase": null
 }
