+++ mcp_server/state/github_read_models.py
@@ -1,49 +1,10 @@
-"""Read models for GitHub API responses."""
+"""Read models for GitHub API responses.
+
+Deprecated: Moved to mcp_server.schemas.github_models.
+"""
 
 from __future__ import annotations
 
-from pydantic import BaseModel, ConfigDict
+from mcp_server.schemas.github_models import IssueReadModel, MilestoneReadModel, PRReadModel
 
-
-class PRReadModel(BaseModel):
-    """Immutable read model for a GitHub pull request."""
-
-    model_config = ConfigDict(frozen=True, extra="forbid")
-
-    pr_number: int
-    title: str
-    state: str
-    base_branch: str
-    head_branch: str
-    merged_at: str | None
-    merge_sha: str | None
-    body: str
-
-
-class MilestoneReadModel(BaseModel):
-    """Immutable read model for a GitHub milestone."""
-
-    model_config = ConfigDict(frozen=True, extra="forbid")
-
-    number: int
-    title: str
-    state: str
-
-
-class IssueReadModel(BaseModel):
-    """Immutable read model for a GitHub issue."""
-
-    model_config = ConfigDict(frozen=True, extra="forbid")
-
-    number: int
-    url: str
-    title: str
-    body: str
-    state: str
-    labels: list[str]
-    milestone: MilestoneReadModel | None
-    assignees: list[str]
-    created_at: str
-    updated_at: str
-    closed_at: str | None
-    author: str
+__all__ = ["IssueReadModel", "MilestoneReadModel", "PRReadModel"]
