+++ tests/mcp_server/unit/managers/test_github_manager.py
@@ -19,7 +19,7 @@ from mcp_server.schemas import (
     MilestoneConfig,
     ScopeConfig,
 )
-from mcp_server.state.github_read_models import IssueReadModel, MilestoneReadModel, PRReadModel
+from mcp_server.schemas.github_models import IssueReadModel, MilestoneReadModel, PRReadModel
 
 
 class TestGitHubManager:
